requirements.txt.txt
protobuf==4.25.3
numpy==1.26.4
mediapipe==0.10.14
opencv-python==4.10.0.84
insightface==0.7.3
google-cloud-pubsub==2.25.0
spacy==3.8.5
grpcio-status==1.62.2
thinc==8.3.4
onnxruntime==1.20.0
scikit-learn==1.3.2
seaborn==0.13.2
matplotlib==3.9.2